feat(core): always replace old ccd with new on dq
refs. metriport/metriport-internal#1730
