--- packages/core/src/external/carequality/dq/find-document-reference.ts
@@ -1,30 +1,0 @@
-import { Config } from "../../../util/config";
-import { S3Utils } from "../../aws/s3";
-import { XDSUnknownPatientId } from "../error";
-import { extractPatientUniqueId } from "../shared";
-const medicalDocumentsBucketName = Config.getMedicalDocumentsBucketName();
-const region = Config.getAWSRegion();
-
-export function decodePatientId(patientIdB64: string): { cxId: string; id: string } | undefined {
-  try {
-    const decodedString = extractPatientUniqueId(patientIdB64);
-    const [cxId, id] = decodedString.split("/");
-
-    if (!cxId || !id) {
-      throw new XDSUnknownPatientId("Patient ID is not valid");
-    }
-    return { cxId, id };
-  } catch (error) {
-    throw new XDSUnknownPatientId("Patient ID is not valid");
-  }
-}
-
-export async function findDocumentReferences(cxId: string, patientId: string): Promise<string[]> {
-  const s3Utils = new S3Utils(region);
-  const documentContents = await s3Utils.retrieveDocumentIdsFromS3(
-    cxId,
-    patientId,
-    medicalDocumentsBucketName
-  );
-  return documentContents;
-}
